Move BaseMessageEventUtil to trace package.
