+++ api/src/test/java/io/opencensus/trace/BaseMessageEventUtilTest.java
@@ -14,12 +14,10 @@
  * limitations under the License.
  */
 
-package io.opencensus.internal;
+package io.opencensus.trace;
 
 import static com.google.common.truth.Truth.assertThat;
 
-import io.opencensus.trace.MessageEvent;
-import io.opencensus.trace.NetworkEvent;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
